Rename `third_party` to `assets`
