+++ packages/extension/src/manifest.json
@@ -37,7 +37,7 @@
   },
   "web_accessible_resources": [
     {
-      "resources": ["third_party/data/*.json", "data/*json"],
+      "resources": ["assets/data/*.json", "data/*json"],
       "matches": ["*://*/*"]
     }
   ]
