+++ packages/extension/src/utils/fetchCookieDictionary.ts
@@ -23,9 +23,7 @@ import { type CookieDatabase } from '@ps-analysis-tool/common';
  * @returns {Promise<CookieDatabase>} Open Cookie Data base
  */
 export async function fetchDictionary(): Promise<CookieDatabase> {
-  const url = chrome.runtime.getURL(
-    'third_party/data/open-cookie-database.json'
-  );
+  const url = chrome.runtime.getURL('assets/data/open-cookie-database.json');
 
   const data = await (await fetch(url)).json();
