+++ packages/extension/src/view/devtools/components/settings/components/settingsContainer.tsx
@@ -25,7 +25,7 @@ import SettingOption from './settingOption';
 import { useSettingsStore } from '../../../stateProviders/syncSettingsStore';
 // @ts-ignore
 // eslint-disable-next-line import/no-relative-packages
-import Gear from '../../../../../../../../third_party/icons/gear.svg';
+import Gear from '../../../../../../../../assets/icons/gear.svg';
 import { SETTING_PAGE_CONTROLS } from '../../../../../constants';
 
 interface settingsToReturnObject {
